bug fixed for create post
